File renames Auth -> Cryptography
File moved to Cryptography -> RC4Engine.h
