+++ src/logonserver/Server/AccountMgr.cpp
@@ -6,7 +6,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "AccountMgr.h"
 #include <Logging/Logger.hpp>
 #include <Logging/Log.hpp>
-#include <Auth/BigNumber.h>
+#include <Cryptography/BigNumber.h>
 #include <Utilities/Strings.hpp>
 #include <Database/Database.h>
 #include "Master.hpp"
