+++ src/shared/CMakeLists.txt
@@ -3,8 +3,8 @@
 # set up our project name
 project(shared CXX)
 
-include(Auth/CMakeLists.txt)
 include(Config/CMakeLists.txt)
+include(Cryptography/CMakeLists.txt)
 include(Database/CMakeLists.txt)
 include(Debugging/CMakeLists.txt)
 include(Exceptions/CMakeLists.txt)
@@ -14,8 +14,8 @@ include(Threading/CMakeLists.txt)
 include(Utilities/CMakeLists.txt)
 
 set(sources
-    ${SRC_AUTH_FILES}
     ${SRC_CONFIG_FILES}
+    ${SRC_CRYPTOGRAPHY_FILES}
     ${SRC_DATABASE_FILES}
     ${SRC_DEBUGGING_FILES}
     ${SRC_EXCEPTIONS_FILES}
@@ -50,7 +50,6 @@ set(headers
     LogonCommDefines.h
     PerformanceCounter.hpp
     PreallocatedQueue.h
-    RC4Engine.h
     SysInfo.hpp
     WorldPacket.h
     WoWGuid.h
