+++ src/world/Server/LogonCommClient/LogonCommHandler.cpp
@@ -8,7 +8,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Server/MainServerDefines.h"
 #include "Server/Master.h"
 #include "Config/Config.h"
-#include "Auth/Sha1.h"
+#include "Cryptography/Sha1.h"
 #include "Server/World.h"
 #include "Utilities/Strings.hpp"
