+++ src/world/Server/WorldSocket.h
@@ -22,7 +22,7 @@
 #define WORLDSOCKET_H
 
 #include "FastQueue.h"
-#include "Auth/WowCrypt.hpp"
+#include "Cryptography/WowCrypt.hpp"
 #include "WorldPacket.h"
 #include "Network/Network.h"
 #include "WorldConf.h"
